+++ commands/waveforms/remove.ts
@@ -8,7 +8,9 @@
            
             try {
                 
+                const b_channel = message.guild?.me?.voice.channel;
                 const u_channel = message.member?.voice?.channel;
+
                 if (!u_channel) return message.channel.send('> You must be in a voice channel to use this command.')
                     .then(message => { setTimeout(() => { message.delete() }, 5000) });
 
@@ -20,35 +22,37 @@
                         .setColor(colors.crimson as ColorResolvable);
                     return message.channel.send({ embeds: [warn] })
                         .then(message => { setTimeout(() => { message.delete() }, 5000) });
-                }
 
-                let merged_queue = [...queue.previousSongs, ...queue.songs];
-                let current_track = (merged_queue.indexOf(queue.songs[0]));
-                let track_index = args[0] - 1 ? args[0] - 1 : -1;
+                } else if (u_channel === b_channel) {
 
-                if (isNaN(args[0]) || args[0] > merged_queue.length || args[0] < 0 || track_index === current_track) {
+                    let merged_queue = [...queue.previousSongs, ...queue.songs];
+                    let current_track = (merged_queue.indexOf(queue.songs[0]));
+                    let track_index = args[0] - 1 ? args[0] - 1 : -1;
 
-                    const warn = new Discord.MessageEmbed()
-                        .setDescription("\`🏴\` ⟶ Invalid arguments.")
-                        .setColor(colors.crimson as ColorResolvable);
-                    return message.channel.send({ embeds: [warn] })
-                        .then(message => { setTimeout(() => { message.delete() }, 5000) });
-            
-                } else {
-                    
-                    let track = merged_queue[track_index];            
-                    if (args[0] > current_track) {
-                        queue.songs.splice(args[0]-1-current_track, 1);
-                    } else if (args[0] < current_track) {
-                        queue.previousSongs.splice(track_index, 1);
+                    if (isNaN(args[0]) || args[0] > merged_queue.length || args[0] < 0 || track_index === current_track) {
+
+                        const warn = new Discord.MessageEmbed()
+                            .setDescription("\`🏴\` ⟶ Invalid arguments.")
+                            .setColor(colors.crimson as ColorResolvable);
+                        return message.channel.send({ embeds: [warn] })
+                            .then(message => { setTimeout(() => { message.delete() }, 5000) });
+                
+                    } else {
+                        
+                        let track = merged_queue[track_index];            
+                        if (args[0] > current_track) {
+                            queue.songs.splice(args[0]-1-current_track, 1);
+                        } else if (args[0] < current_track) {
+                            queue.previousSongs.splice(track_index, 1);
+                        }
+
+                        const main = new Discord.MessageEmbed()
+                            .setDescription(`✦ Removed track [${track.name}](${track.url})`)
+                            .setColor(colors.crimson as ColorResolvable)
+                            .setFooter({ text: `Arkus.wav  •  Removed by ${message.author.username}   ` })
+                            .setTimestamp();
+                        return message.channel.send({ embeds: [main] });
                     }
-
-                    const main = new Discord.MessageEmbed()
-                        .setDescription(`✦ Removed track [${track.name}](${track.url})`)
-                        .setColor(colors.crimson as ColorResolvable)
-                        .setFooter({ text: `Arkus.wav  •  Removed by ${message.author.username}   ` })
-                        .setTimestamp();
-                    return message.channel.send({ embeds: [main] });
                 }
             
             } catch(err) {
